requirements.txt
fastapi
uvicorn
flask

gradio

langchain
langchain-community
langchain-huggingface
langchain-ollama
langchain-openai
langchain-qdrant
langchain-text-splitters

langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk

openai
ollama

transformers
torch
sentence-transformers
huggingface_hub
tiktoken

qdrant-client
redis

pandas
numpy
scikit-learn
scipy

pypdf
beautifulsoup4
aiohttp
requests

pydantic
pydantic-settings

sqlalchemy

python-dotenv

notebook
jupyterlab
ipykernel
ipywidgets

onnxruntime

pytest
ruff